% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forNowTransforms.R
\name{applyInsert}
\alias{applyInsert}
\title{Given an array of values and elements, calculates some summary statistics
of some columns}
\usage{
applyInsert(vec, var_cats, a_func)
}
\arguments{
\item{vec}{An array of values and elements}

\item{var_cats}{The `Categories` object that corresponds to the object `vec`}

\item{a_func}{A function usually to calculates a summary statistic}
}
\value{
If `vec` has three dimentions, a row that results from applying `a_func`
 to each column of each object of the third dimention of `vec`; otherwise,
 a row that results from applying `a_func` to each column of `vec`
}
\description{
Given an array of values and elements, calculates some summary statistics
of some columns
}
