% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{applyLatexStyle}
\alias{applyLatexStyle}
\title{Font decoration}
\usage{
applyLatexStyle(item, item_theme)
}
\arguments{
\item{item}{The text to wrap in font decoration}

\item{item_theme}{The font decorations to apply}
}
\description{
Apply font decorations like:
\itemize{
\item underline / underline2 \link{underline}
\item italics from \link{italics}
\item bold from \link{bold}
\item fontsize \link{fontsize}
\item color defined inside this function
}
}
