% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{as.data.frame.CategoricalVariable}
\alias{as.data.frame.CategoricalVariable}
\title{Categorical Variable to data.frame}
\usage{
\method{as.data.frame}{CategoricalVariable}(x, ...)
}
\arguments{
\item{x}{A CategoricalVariable from a crunch \link[crunch]{loadDataset}}

\item{...}{Ignored}
}
\description{
Manipulate categorical into a results object without hitting the tabBook
endpoint. For the purpose of creating a codebookItem.
}
