% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asToplineCategoricalArray.R
\name{as.ToplineCategoricalArray}
\alias{as.ToplineCategoricalArray}
\title{Combine two or more tabbook results as a categorical array}
\usage{
as.ToplineCategoricalArray(questions, question_alias = NULL, labels)
}
\arguments{
\item{questions}{A list of two or more results objects. Objects should be
provided in time or
wave 3, ..., wave n)}

\item{question_alias}{The name of the question alias across all results 
sets. It must be the same in all datasets.}

\item{labels}{A character vector of labels that will be displayed in the
resulting crunchtabs output. Should match the number of results objects}
}
\description{
Here we manipulate the tabBook results so that they match the layout
of a categoricalArray, which has the benefit of already having
distinct code to write it to latex.
}
