% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{banner}
\alias{banner}
\title{Create an object of class \code{Banner}}
\usage{
banner(dataset, vars, labels = NULL, recodes = NULL)
}
\arguments{
\item{dataset}{A Crunch dataset.}

\item{vars}{A list of vectors of names of the variables that shoulds constitute the banner object.
These should be categorical (and typically hidden) variables present in the dataset.}

\item{labels}{An optional named list of labels, where names are variable aliases
(present in the vars parameter) and values are labels that should be used in the report.
for those variables. Defaults to NULL - variables names are used.}

\item{recodes}{An optional named list of categories recodes (the syntax is similar
to the one used in the \code{recode} function in the R \code{car} library). Use \code{NA} to
exclude categories. Not listed categories will be left unchanged.
Use \code{else} to change all not listed categories. See examples for details.
Defaults to NULL - categories are not modified.}
}
\value{
An object of class \code{Banner}.
}
\description{
Create an object of class \code{Banner}
}
\examples{
\dontrun{
crunch_dataset <- loadDataset('dataset_name')
banner(crunch_dataset, vars = list(banner1 = c('var1', 'var2'), banner2 = c('var3')),
    labels = c(var1 = 'var1 label', var2 = 'var2 label'),
    recodes = list(
        var1 = ''cat1a' = 'new cat1a'; 'cat1b' = NA',
        var2 = ''cat2a' = 'new cat2a'; else = NA'))
}
}
