% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{bannerDataRecode}
\alias{bannerDataRecode}
\title{Banner Data Recode}
\usage{
bannerDataRecode(b_table, b_recode)
}
\arguments{
\item{b_table}{A categorical matrix}

\item{b_recode}{A categorical matrix identifying the new category names}
}
\description{
Recodes the names of categories based
}
