% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_v2.R
\name{bcodeBookItemTxtDescription}
\alias{bcodeBookItemTxtDescription}
\title{Extract basic question information}
\usage{
bcodeBookItemTxtDescription(x, ...)
}
\arguments{
\item{x}{A dataset variable}
}
\description{
Extracts the following:
}
\details{
\itemize{
\item alias
\item description or question text
\item notes or filter text
}
}
