% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_v2.R
\name{bwriteCodeBook}
\alias{bwriteCodeBook}
\title{Create a codebook}
\usage{
bwriteCodeBook(x, ...)
}
\arguments{
\item{...}{Additional arguments. Unused.}

\item{ds}{A crunch dataset}
}
\description{
Create a codebook
}
