% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forNowTransforms.R
\name{calcInsertions}
\alias{calcInsertions}
\title{Given a vector of values and elements, calculate the insertions}
\usage{
calcInsertions(vec, elements, var_cats)
}
\arguments{
\item{vec}{values to transform (a single dimension of an array)}

\item{elements}{AbstractCategories of both `Category`s and `Insertion`s to
calculate. Generally derived from `mapInsertions()`}

\item{var_cats}{the `Categories` object tat corresponds to the vector in
`vec` of the transform}
}
\value{
the values given in `vec`, with any insertions specified in
`trans` calculated and inserted
}
\description{
Given a vector of values and elements, calculate the insertions
}
\keyword{internal}
