% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forNowTransforms.R
\name{calcTabMeanInsert}
\alias{calcTabMeanInsert}
\title{Given a vector of values, calculates the weighted mean}
\usage{
calcTabMeanInsert(vec, var_cats)
}
\arguments{
\item{vec}{A vector to use for calculation (a single dimension of an array)}

\item{var_cats}{The `Categories` object that corresponds to the vector in `vec`}
}
\value{
The weighted mean of the numeric values of `var_cats` given the weights
 in `vec`
}
\description{
Given a vector of values, calculates the weighted mean
}
