% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{calculateIfLongtable}
\alias{calculateIfLongtable}
\alias{calculateIfLongtable.default}
\alias{calculateIfLongtable.CrossTabVar}
\alias{calculateIfLongtable.ToplineVar}
\alias{calculateIfLongtable.ToplineCategoricalArray}
\title{Calculate number of lines}
\usage{
calculateIfLongtable(x, theme)

\method{calculateIfLongtable}{default}(x, theme)

\method{calculateIfLongtable}{CrossTabVar}(x, theme)

\method{calculateIfLongtable}{ToplineVar}(x, theme)

\method{calculateIfLongtable}{ToplineCategoricalArray}(x, theme)
}
\arguments{
\item{x}{An object of type: CrossTabVar, ToplineVar, or ToplineCategoricalArray}

\item{theme}{An object created by \link{themeNew}}
}
\description{
Calculates the length of the table depending on
the type of data provided as well as some theme
information
}
