% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catArrayTexAdjustments.R
\name{catArrayTexAdjustments}
\alias{catArrayTexAdjustments}
\title{categoricalArray tex adjustments}
\usage{
catArrayTexAdjustments(
  result,
  category_aliases = NULL,
  category_text = NULL,
  width = NULL,
  bullets = TRUE
)
}
\arguments{
\item{result}{A crosstab result sub-object that is typically passed to \code{\link{writeLatex}}}

\item{category_text}{Defaults to NULL. Otherwise, overwrites the category text}

\item{width}{A number representing the width of the LaTeX parbox that contains the category
list}

\item{bullets}{A logical. Defaults to TRUE. Should the output have bullets?}

\item{category_alias}{Defaults to NULL. Otherwise, the categories will be itemized using
item_alias.}
}
\description{
If a user wishes to show category/statement data that is too large to be shown comfortably on
the width of the PDF document, we provide functionality for renaming the columns headers of a
categoricalArray's presentation and putting them above or below the table in a itemized list.
}
\details{
The default itemization scheme is uppercase letters. If you do not provide shortnames or aliases
(not to be confused with crunch aliases), you will see A, B, C, and so on for each column of
your categorical array
}
