% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoryCheck.R
\name{categoryCheck}
\alias{categoryCheck}
\title{Verify category stability}
\usage{
categoryCheck(dataset_list, var, warnings_only = FALSE)
}
\arguments{
\item{dataset_list}{A list of datasets in which the variable resides}

\item{var}{The variable to test}

\item{warnings_only}{Should we present a warning or an error?}
}
\description{
This function allows one to review a group of questions and identify if they can be reasonably
joined together for the purpose of merging, appending, or creating a tracking report.
}
