% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoryCheck.R
\name{categoryCheckAvailable}
\alias{categoryCheckAvailable}
\title{Check alias availability}
\usage{
categoryCheckAvailable(dataset_list, var, warnings_only = FALSE)
}
\arguments{
\item{dataset_list}{A list of datasets in which the variable resides}

\item{var}{The variable to test}

\item{warnings_only}{Should we present a warning or an error?}
}
\description{
Given a list of datasets, check which datasets the alias is available inside and present
a warning or error for those that it is not available in.
}
