% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoryCheck.R
\name{categoryCheckConsistency}
\alias{categoryCheckConsistency}
\title{Check category consistency}
\usage{
categoryCheckConsistency(results_list, var, warnings_only = FALSE)
}
\arguments{
\item{results_list}{A list of tabBook results from \link{\code{trackingRepotrs_tabs}}}

\item{var}{The variable to test}

\item{warnings_only}{Should we present a warning or an error?}

\item{fix}{Should we expand categories with zeros so that they do not cause failures
when writing to LaTex or Excel?}

\item{output}{Should we return a list of categories in each of the results s}
}
\description{
Given a list of results from tracking reports. Check the consistency of aliases and optionally
expand (0 filling) the categories so that we can proceed to writing to PDF or Excel
}
