% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asToplineCategoricalArray.R
\name{categoryFill}
\alias{categoryFill}
\title{Column Bind Unequal Matrices}
\usage{
categoryFill(clist)
}
\arguments{
\item{clist}{A list of matrices with rownames}
}
\description{
This function takes a list of matrices and binds them together into
a single frame. Accounts for missing or unequal rows, by rowname.
}
