% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asToplineCategoricalArray.R
\name{cbindFill}
\alias{cbindFill}
\title{Merge two data.frames by rownames}
\usage{
cbindFill(x, y)
}
\arguments{
\item{x}{A data.frame}

\item{y}{A data.frame}
}
\description{
This function is designed to cbind via rownames where
the rownames may not match and then fix the result so that
it can be further merged to another data.frame.
}
