% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{cmidrule}
\alias{cmidrule}
\title{cmidrule tex}
\usage{
cmidrule(from, to, size = "lr{.75em}")
}
\arguments{
\item{from}{The column number}

\item{to}{The column number}

\item{size}{The size of the rule (line)}
}
\description{
Used for creating rule (lines) underneath a sub-header in
tables
}
