% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codeBookItem}
\alias{codeBookItem}
\title{Generate a codebook item}
\usage{
codeBookItem(x, ...)
}
\arguments{
\item{x}{An object of one of the types listed}

\item{...}{Additional arguents passed to codebookItem methods}
}
\description{
A passthrough function that creates a table header appropriate to the class
of the data object being passed. Expected classes are:
}
\details{
\itemize{
\item CategoricalVariable
\item CategoricalArrayVariable
\item MultipleResponseVariable
\item NumericVariable
\item TextVariable
\item DatetimeVariable
}

Importantly, this also controls the relative widths of the columns.
}
