% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebookItemTxt}
\alias{codebookItemTxt}
\title{Extract basic question information}
\usage{
codebookItemTxt(x)
}
\arguments{
\item{x}{A dataset variable}
}
\description{
Extracts the following:
}
\details{
\itemize{
\item body
\item alias
\item description or question text
\item notes or filter text
\item id
}
}
