% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookLatex.R
\name{codeBookItemTxtDescription}
\alias{codeBookItemTxtDescription}
\title{Extract basic question information}
\usage{
codeBookItemTxtDescription(x, ...)
}
\arguments{
\item{x}{A dataset variable}

\item{...}{Additional arguments passed to \link{kable_styling_defaults}}
}
\description{
Extracts the following:
}
\details{
\itemize{
\item alias
\item description or question text
\item notes or filter text
}
}
