% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookGeneric.R
\name{codeBookItemTxtDescriptionGeneral}
\alias{codeBookItemTxtDescriptionGeneral}
\title{Extract basic question information}
\usage{
codeBookItemTxtDescriptionGeneral(x, nm, meta, ...)
}
\arguments{
\item{x}{A dataset variable}

\item{nm}{The alias of the current codeBookItem}

\item{meta}{A specifically formatted data.frame with meta data.}

\item{...}{Additional arguments passed to \link{kable_styling_defaults}}
}
\description{
Extracts the following:
}
\details{
\itemize{
\item alias
\item description or question text
\item notes or filter text
}
}
