% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codeBookItemTxtHeader}
\alias{codeBookItemTxtHeader}
\title{codeBook Item Text Header}
\usage{
codeBookItemTxtHeader(x, ...)
}
\arguments{
\item{x}{A dataset variable}

\item{...}{Additional arguments passed to \link{kable_styling_defaults}}
}
\description{
Creates a text header for a codebook item
}
