% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookGeneric.R
\name{codeBookItemTxtHeaderGeneral}
\alias{codeBookItemTxtHeaderGeneral}
\title{codeBook Item Text Header}
\usage{
codeBookItemTxtHeaderGeneral(x, nm, meta, ...)
}
\arguments{
\item{x}{An R vector}

\item{nm}{The alias of the current codeBookItem}

\item{meta}{A specifically formatted data.frame with meta data.}
}
\description{
Creates a text header for a codebook item
}
