% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_items}
\alias{collapse_items}
\title{Collapse items}
\usage{
collapse_items(x, and = FALSE, or = FALSE, quotes = FALSE)
}
\arguments{
\item{x}{A character string}

\item{and}{Logical, collapse using "and"}

\item{or}{Logical, collapse using "or"}

\item{quotes}{Logical, should it be quoted?}
}
\description{
Rule based collapse
}
