% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{compute_pvals}
\alias{compute_pvals}
\title{Column based hypothesis testing}
\usage{
compute_pvals(counts, counts_unweighted)
}
\arguments{
\item{counts}{A data.frame or matrix of counts}

\item{counts_unweighted}{A data.frame or matrix of counts}
}
\description{
Calculates chi-square and returns p-values
}
