% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabs.R
\name{crosstabs}
\alias{crosstabs}
\alias{toplines}
\title{Toplines and Crosstabs Summaries}
\usage{
crosstabs(
  dataset,
  vars = names(dataset),
  weight = crunch::weight(dataset),
  banner = NULL,
  codebook = FALSE,
  include_numeric = FALSE,
  include_datetime = FALSE,
  include_verbatims = FALSE,
  num_verbatims = 10
)

toplines(
  dataset,
  vars = names(dataset),
  weight = crunch::weight(dataset),
  banner = NULL,
  codebook = FALSE,
  include_numeric = FALSE,
  include_datetime = FALSE,
  include_verbatims = FALSE,
  num_verbatims = 10
)
}
\arguments{
\item{dataset}{A Crunch dataset.}

\item{vars}{An optional vector of aliases of the non-hidden variables that shoulds be used.
Defaults to all non-hidden variables.}

\item{weight}{The alias of a numeric variable that should be used for data weighting.
Defaults to current weight variable. For unweighted, set to \code{NULL}}

\item{banner}{An optional object of class \code{Banner} that should be used to generate
a Crosstabs summary. Defaults to \code{NULL} - a Toplines summary is produced and returned.}

\item{codebook}{If \code{TRUE}, codebook data summaries are prepared. Defaults to \code{FALSE}.}

\item{include_numeric}{Logical. Should we include numeric questions? Defaults to FALSE. Implemented for Toplines only.}

\item{include_datetime}{Logical. Should we include date time questions? Defaults to FALSE. Implemented for Toplines only.}

\item{include_verbatims}{Logical. Should we include a sample text varaibles? Defaults to FALSE. Implemented for Toplines only.}

\item{num_verbatims}{An integer identifying the number of examples to extract from a text variable. Defaults to 10. Implemented for Toplines only.}
}
\value{
A Toplines (when no banner is provided) or Crosstabs (when a banner is provided)
summary of the input dataset.
}
\description{
\code{crosstabs} produces Toplines (one-way frequency tables) or Crosstabs (cross tabulations)
summaries of a Cruch dataset.
}
\section{Functions}{
\itemize{
\item \code{toplines}: An alias for \code{crosstabs}
}}

\examples{
\dontrun{
toplines_summary <- crosstabs(crunch_dataset, weight = 'weight')
crosstabs_summary <- crosstabs(crunch_dataset, vars = c('alias1', 'alias2'),
 weight = 'weight', banner = banner_object)
}
}
