% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchtabs.R
\docType{package}
\name{crunchtabs-package}
\alias{crunchtabs}
\alias{crunchtabs-package}
\title{Functions for reports (toplines and banners) generation.}
\description{
crunchtabs provides tools for generating reports (toplines and banners)
}

