% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_if_items}
\alias{error_if_items}
\title{error if items}
\usage{
error_if_items(
  items,
  text,
  error = TRUE,
  and = FALSE,
  or = FALSE,
  quotes = FALSE
)
}
\arguments{
\item{items}{A character vector of items}

\item{text}{A character vector that items may be within}

\item{error}{Logical, should a stop be called?}

\item{and}{Logical, passed to \link{collapse_items}}

\item{or}{Logical, passed to \link{collapse_items}}

\item{quotes}{Logical, passed to \link{collapse_items}}
}
\description{
A function assertion used to
create an error condition on the existence
of some items
}
