% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{fixExtraSummaryWeights}
\alias{fixExtraSummaryWeights}
\title{Weighting Object Titles}
\usage{
fixExtraSummaryWeights(obj)
}
\arguments{
\item{obj}{The obj created by \link{prepareExtraSummary}}
}
\description{
If the object is weighted
there is a slight difference in
the naming required to align
with existing patterns
}
