% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{formatNum}
\alias{formatNum}
\title{Format Number}
\usage{
formatNum(x, digits = 0, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{The number of digits}

\item{...}{Furth arguments, unused.}
}
\description{
Round and format a number, removing what space.
}
