% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookGeneric.R
\name{get_class}
\alias{get_class}
\title{Get class name from arrow variable}
\usage{
get_class(ds, variable)
}
\arguments{
\item{ds}{An arrow dataset}

\item{variable}{A string identifying the variable}
}
\description{
Arrow does not provide a method for class on it's variables. This function prints out
the schema and captures the appropriate output.
}
