% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{get_grid_number}
\alias{get_grid_number}
\title{Return Excel-style column name.}
\usage{
get_grid_number(n)
}
\arguments{
\item{n}{An integer identfying the column number}
}
\description{
Returns an excel style column name. Useful
for identifying a specific column. Column 1, would be A
in Excel, column 100 would be CV in excel
}
