% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catArrayTexAdjustments.R
\name{getCatArrayColNames}
\alias{getCatArrayColNames}
\title{getter for Array column names}
\usage{
getCatArrayColNames(result)
}
\arguments{
\item{result}{A crosstab result sub-object that is typically passed to \code{\link{writeLatex}}}
}
\description{
getter for Array column names
}
