% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getCounts}
\alias{getCounts}
\title{Extract counts from a tabbook}
\usage{
getCounts(x, alias)
}
\arguments{
\item{x}{A results object or element}

\item{alias}{The alias that you wish to extract counts for}
}
\description{
Extract counts from a tabbook
}
