% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{getItemData}
\alias{getItemData}
\title{get Item Data for print}
\usage{
getItemData(data, item_name, empty_col, round)
}
\arguments{
\item{data}{A list or data frame}

\item{item_name}{A string identifying the item}

\item{empty_col}{A logical identifying if the column is empty}

\item{round}{A logical identifying if the item should be rounded}
}
\description{
Retrieves item data and prepares it for printing
}
