% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{getMultitable}
\alias{getMultitable}
\title{Get Multitable}
\usage{
getMultitable(banner_flatten, dataset)
}
\arguments{
\item{banner_flatten}{A banner object from \link{banner}}

\item{dataset}{The name of the data}
}
\description{
Given a Banner object and a dataset, find or create the corresponding Crunch multitable
}
