% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getName}
\alias{getName}
\alias{getName.default}
\alias{getName.ToplineBase}
\alias{getName.BannerVar}
\alias{getName.CrossTabVar}
\alias{getName.CrunchCube}
\alias{getName.CrunchTabs}
\title{S3 Method for getName}
\usage{
getName(x)

\method{getName}{default}(x)

\method{getName}{ToplineBase}(x)

\method{getName}{BannerVar}(x)

\method{getName}{CrossTabVar}(x)

\method{getName}{CrunchCube}(x)

\method{getName}{CrunchTabs}(x)
}
\arguments{
\item{x}{An object of class CrunchTabs}
}
\description{
getName returns....
}
