% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getProportions}
\alias{getProportions}
\title{Extract proportions from a tabbook}
\usage{
getProportions(x, alias)
}
\arguments{
\item{x}{A results object or element}

\item{alias}{The alias that you wish to extract proportions for}
}
\description{
Extract proportions from a tabbook
}
