% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{getResultObj}
\alias{getResultObj}
\title{Generic getter for result object}
\usage{
getResultObj(results, alias)
}
\arguments{
\item{results}{A list of results}

\item{alias}{A character identifying the alias}

\item{obj}{A result object}
}
\description{
Get a result object based on alias
}
