% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{getVarInfo}
\alias{getVarInfo}
\title{Get Variable Info}
\usage{
getVarInfo(var, theme)
}
\arguments{
\item{var}{The crunch variable}

\item{theme}{The theme object from \link{themeNew}}
}
\description{
Given a crunch variable and a theme object
prepare a list of meta data to package or
subset from the variable for use in the
application of the theme.
}
\details{
\itemize{
\item alias
\item name
\item description
\item filtertext
\item subname
}
}
