% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifelseOverload}
\alias{ifelseOverload}
\alias{\%||\%}
\title{Operator overload}
\usage{
a \%||\% b
}
\arguments{
\item{a}{An R object}

\item{b}{An R object}
}
\description{
Creates an ifelse function that handles null
}
