% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebookLatex.R
\name{kable_strip_toprules}
\alias{kable_strip_toprules}
\title{Strip toprule}
\usage{
kable_strip_toprules(x)
}
\arguments{
\item{x}{A character string}
}
\description{
Strip horizontal lines (also called rules) from
codebooks generated for latex
}
