% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{kable_styling_defaults}
\alias{kable_styling_defaults}
\title{Defaults for kableExtra}
\usage{
kable_styling_defaults(x, ...)
}
\arguments{
\item{x}{A kable object}

\item{...}{Additional arguments passed to \link[kableExtra]{kable_styling}}
}
\description{
Default styling for kable extra
}
