% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLatex.R
\name{latexDocHead}
\alias{latexDocHead}
\title{Prepare Doc Head}
\usage{
latexDocHead(theme, title, subtitle, banner = NULL)
}
\arguments{
\item{theme}{A theme object created by \link{themeNew}}

\item{title}{A page title}

\item{subtitle}{A page subtitle}

\item{banner}{A banner object from \link{banner}}
}
\description{
Prepare a document head for a latex document
}
