% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLatex.R
\name{latexSampleDescription}
\alias{latexSampleDescription}
\title{Sample Description}
\usage{
latexSampleDescription(sample_desc, field_period, moe)
}
\arguments{
\item{sample_desc}{A string describing the samepl}

\item{field_period}{A string defining from-to field dates}

\item{moe}{The margin of error for the survey}
}
\description{
Adds a sample description, field window and margin of error to
your result document.
}
