% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{latexTableBody}
\alias{latexTableBody}
\title{Create Tex of Table Body}
\usage{
latexTableBody(df, theme)
}
\arguments{
\item{df}{An object from \link{reformatLatexResults}}

\item{theme}{A theme object from \link{themeNew}}
}
\description{
Prepare the TeX for the table body given initial
content and apply theme elements.
}
