% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{latexTableName}
\alias{latexTableName}
\title{Question Text}
\usage{
latexTableName(var, theme)
}
\arguments{
\item{var}{A crunch variable}

\item{theme}{An object created by \link{themeNew}}
}
\description{
Creates the question text that leads a topline or crosstab
}
