% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{longtableHeadFootMacros}
\alias{longtableHeadFootMacros}
\title{Generates two macros for the preamble
 \\bannera{} that takes one argument (first column label)
 \\tbltopa that takes no arguments
 If given multiple banners, \\bannerb \\tbltopb, etc are created}
\usage{
longtableHeadFootMacros(banner, num, page_width = 9, theme)
}
\arguments{
\item{banner}{A banner object from \link{banner}}

\item{num}{An integer that later specifies a letter based on the integers intrinsic position}

\item{page_width}{The width of the page in inches (typically 9 landscape or 6.5 portrait)}

\item{theme}{A theme object from \link{themeNew}}
}
\description{
Generates two macros for the preamble
 \\bannera{} that takes one argument (first column label)
 \\tbltopa that takes no arguments
 If given multiple banners, \\bannerb \\tbltopb, etc are created
}
