% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{makeLatexBanner}
\alias{makeLatexBanner}
\title{Make Banner}
\usage{
makeLatexBanner(binfo, width = NULL)
}
\arguments{
\item{binfo}{banner info from \link{getBannerInfo}}

\item{width}{Unused argument}
}
\description{
Provides spacing and additional formatting for
LaTeX banners.
}
