% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookLatex.R
\name{monospaced}
\alias{monospaced}
\title{Fix ttf}
\usage{
monospaced(x)
}
\arguments{
\item{x}{A string containing an underscore}
}
\description{
We must escape underscores in aliases because latex treats them
like mathematical subtext
}
