% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{multicolumn}
\alias{multicolumn}
\title{multicolumn tex}
\usage{
multicolumn(width = 1, ..., align = "c")
}
\arguments{
\item{width}{The number of columns}

\item{...}{Further arguments passed to multicolumn}

\item{align}{One of l, c or r.}
}
\description{
Creates a programmable multicolumn with width and alignment
}
