% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{newcommand}
\alias{newcommand}
\title{newcommand tex}
\usage{
newcommand(command, ..., args = 0)
}
\arguments{
\item{command}{Numeric value in first brackets.}

\item{...}{Numeric value in second brackets.}

\item{args}{Numeric value in square brackets.}
}
\description{
Creates a newcommand tex string
}
