% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookLatex.R
\name{noBreaks}
\alias{noBreaks}
\title{No breaks allowed}
\usage{
noBreaks(tex)
}
\arguments{
\item{tex}{A string including escaped tex}
}
\description{
This code wraps arbitrary tex inside of a block that
will not break until it is longer than a page
}
