% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{nonTabBookSummary.NumericVariable}
\alias{nonTabBookSummary.NumericVariable}
\title{Prepare Numeric Content}
\usage{
\method{nonTabBookSummary}{NumericVariable}(x, weighted = TRUE, num = 10, tz = "UTC", ...)
}
\arguments{
\item{x}{A variable of class \link[crunch]{NumericVariable}}

\item{weighted}{Logical. Are these data weighted?}

\item{num}{The number of verbatim responses to present as a sample. Defaults to 10.}

\item{tz}{A timezone. Defaults to UTC.}

\item{...}{Additional arguments passed to methods}
}
\description{
\link[crunch]{tabBook} does not report an appropriate numeric summary
without being provided with a multitable. So we "fake" a numeric summary
by overwriting the structure of a categorical object.
}
\details{
If data are weighted we display Weighted N instead of Unweighted
N
}
