% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_around}
\alias{paste_around}
\title{Paste Around}
\usage{
paste_around(str, before, after)
}
\arguments{
\item{str}{A character string}

\item{before}{A character string to go before str}

\item{after}{A character string to go after str}
}
\description{
Embeds a string in given before and after
}
