% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdflatex}
\alias{pdflatex}
\title{pdflatex}
\usage{
pdflatex(
  texfile,
  open = interactive(),
  verbose = FALSE,
  cleanup = TRUE,
  options = "-halt-on-error",
  path.to.pdflatex = Sys.which("pdflatex"),
  ...
)
}
\arguments{
\item{texfile}{The path to the texfile}

\item{open}{Logical, should the result be opened?}

\item{verbose}{Logical, should the function be verbose?}

\item{cleanup}{Logical, should tex/log files be cleaned?}

\item{options}{Additional options passed to pdflatex}

\item{path.to.pdflatex}{The system path to pdflatex}
}
\description{
A utility function that runs pdflatex
}
