% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{prepareExtraSummary.DatetimeVariable}
\alias{prepareExtraSummary.DatetimeVariable}
\title{Prepare Datetime Content}
\usage{
\method{prepareExtraSummary}{DatetimeVariable}(x, weighted = TRUE, tz = "UTC")
}
\arguments{
\item{x}{A variable of class \link[crunch]{DatetimeVariable}}

\item{weighted}{Logical. Are these data weighted?}

\item{tz}{A timezone. Defaults to UTC.}
}
\description{
tabBook does not report an appropriate date time summary without
being provided with a multitable. So we "fake" a date time summary
by overwriting the structure of a categorical object.
}
\details{
If data are weighted we display Weighted N instead of Unweighted
N
}
