% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{prepareExtraSummary.NumericVariable}
\alias{prepareExtraSummary.NumericVariable}
\title{Prepare Numeric Content}
\usage{
\method{prepareExtraSummary}{NumericVariable}(x, weighted = TRUE)
}
\arguments{
\item{x}{A variable of class \link[crunch]{NumericVariable}}

\item{weighted}{Logical. Are these data weighted?}
}
\description{
tabBook does not report an appopriate numeric summary without
being provided with a multitable. So we "fake" a numeric summary
as by overwriting the structure of a categorical object.
}
\details{
If data are weighted we display Weighted N instead of Unweighted
N
}
