% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{prepareExtraSummary}
\alias{prepareExtraSummary}
\alias{prepareExtraSummary.default}
\title{Prepare Summary Content}
\usage{
prepareExtraSummary(x, ...)

\method{prepareExtraSummary}{default}(x)
}
\arguments{
\item{x}{A variable of class NumericVariable, DatetimeVariable or TextVariable}

\item{...}{Additional arguments passed to methods}
}
\description{
Prepare summary content for toplines for classes that
are not covered by tabBook such as NumericVariable, DatetimeVariables
and TextVariable
}
