% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{prepareExtraSummary.TextVariable}
\alias{prepareExtraSummary.TextVariable}
\title{Prepare Text Content}
\usage{
\method{prepareExtraSummary}{TextVariable}(x, weighted = TRUE, num = 10L)
}
\arguments{
\item{x}{A variable of class \link[crunch]{TextVariable}}

\item{weighted}{Logical. Are these data weighted?}

\item{num}{The number of verbatim responses to present as a sample. Defaults to 10.}
}
\description{
tabBook does not report an appropriate date time summary without
being provided with a multitable. So we "fake" a date time summary
by overwriting the structure of a categorical object.
}
\details{
If data are weighted we display Weighted N instead of Unweighted
N
}
