% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{print.Crosstabs}
\alias{print.Crosstabs}
\title{Crosstabs Title}
\usage{
\method{print}{Crosstabs}(x, ...)
}
\arguments{
\item{x}{Variable to get name from.}

\item{...}{Further arguments. Unused.}
}
\description{
print.Crosstabs returns the title...
}
