% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{print.Toplines}
\alias{print.Toplines}
\title{Topline Title}
\usage{
\method{print}{Toplines}(x, ...)
}
\arguments{
\item{x}{Variable to get name from.}

\item{...}{Further arguments. Unused.}
}
\description{
print.Toplines returns the title...
}
