% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{project-themes}
\alias{project-themes}
\alias{themeUKPolitical}
\alias{themeHuffPoToplines}
\alias{themeHuffPoCrosstabs}
\title{Project-specific themes}
\usage{
themeUKPolitical()

themeHuffPoToplines(logo = NULL)

themeHuffPoCrosstabs(logo = NULL)
}
\arguments{
\item{logo}{an image of the logo to use (default: `NULL`)}
}
\value{
a new theme object
}
\description{
These are some default themes for specific projects
}
