% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{recode_categories}
\alias{recode_categories}
\title{Recode Categories}
\usage{
recode_categories(alias, responses, recodes)
}
\arguments{
\item{alias}{A question alias}

\item{responses}{The responses for the specified question alias}

\item{recodes}{A character vector of recodes for those responses}
}
\description{
Given a question alias, recode the responses in 
that question based on the recodes provided.
}
