% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{reflowQuestionNumbers}
\alias{reflowQuestionNumbers}
\title{Reflow Question Numbers}
\usage{
reflowQuestionNumbers(x)
}
\arguments{
\item{x}{A crosstabs object from the \link{crosstabs} function.}
}
\description{
When we manually add summaries of
numeric, datetime or text variables
we must "reflow" the question numbers
so that they match dataset order.
}
