% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{reformatLatexResults}
\alias{reformatLatexResults}
\title{Apply reformats}
\usage{
reformatLatexResults(result, banner, theme)
}
\arguments{
\item{result}{A summarized crunchtab variable}

\item{banner}{A crunchtabs \link{banner}}

\item{theme}{A crunchtabs theme object from \link{themeNew}}
}
\description{
A wrapper for reformatVar that
loops over each of the banners.
}
