% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{reformatResultsGen}
\alias{reformatResultsGen}
\title{reformatResults <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    UseMethod("reformatResults", x)
}}
\usage{
reformatResultsGen(x, proportions, theme, reformat, details)
}
\description{
#' @export
reformatResults.default <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    stop("`reformatResults` generic function doesn't support objects of type: ",
        collapse_items(class(x)))
}
}
\details{
#' @export
reformatResults.ToplineBase <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}

#' @export
reformatResults.CrossTabVar <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}

#' @export
reformatResults.ToplineNumeric <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}

#' @export
reformatResults.CrossTabBannerVar <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}
}
