% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{reformatVar}
\alias{reformatVar}
\title{Reformat variables for printing}
\usage{
reformatVar(var, banner_name, theme, proportions, banner_info, latex)
}
\arguments{
\item{var}{A crunch variable}

\item{banner_name}{A string identfying the name of the banner}

\item{theme}{A theme object created by \link{themeNew}}

\item{proportions}{Should proportions be displayed or counts? One of "proportions" or "counts"}

\item{banner_info}{A meta data object from \link{getBannerInfo}}

\item{latex}{A logical identifying if this code is for LaTex or Excel. If TRUE, LaTeX}
}
\description{
This function prepares data for placement in tex. By taking the following actions:
}
\details{
\itemize{
\item Adding totals_row
\item Adding weighted/un-weighted base rows
\item Placing table content appropriately
}
}
