% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabelDescription}
\alias{relabelDescription}
\title{relabel description}
\usage{
relabelDescription(obj, labs)
}
\arguments{
\item{obj}{A results object for a specific alias}

\item{labs}{A list of labels, typically passed to \link{relabel}}
}
\description{
Given a results object, adjust the question description. This will display as the question
text in crunchtabs.
}
