% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabelNotes}
\alias{relabelNotes}
\title{relabel notes}
\usage{
relabelNotes(obj, labs)
}
\arguments{
\item{obj}{A results object for a specific alias}

\item{labs}{A list of labels, typically passed to \link{\code{relabel}}}
}
\description{
Given a results object, adjust the note. This will display as the text below
the question text in crunchtabs.
}
