% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabelSet}
\alias{relabelSet}
\title{Set relabels}
\usage{
relabelSet(obj, labs)
}
\arguments{
\item{obj}{An inner object under the results list of a \link{\code{crosstabs}} object.}

\item{labs}{A list including at least one of the following: description, notes, items, or options}
}
\description{
An S3 method for applying updated text values to a result object in crunchtabs
}
