% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabelSet.ToplineCategoricalArray}
\alias{relabelSet.ToplineCategoricalArray}
\title{relabel ToplineCategoricalArray}
\usage{
\method{relabelSet}{ToplineCategoricalArray}(obj, labs)
}
\arguments{
\item{obj}{A results object for a specific alias}

\item{labs}{A list of labels, typically passed to \link{\code{relabel}}}
}
\description{
Relabel a ToplineCategoricalArray result object.
}
