% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabelSet.ToplineVar}
\alias{relabelSet.ToplineVar}
\title{relabelSet for ToplineCategoricalArray}
\usage{
\method{relabelSet}{ToplineVar}(obj, labs)
}
\description{
relabelSet for ToplineCategoricalArray
}
