% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabelSet.TopliuneCategoricalArray}
\alias{relabelSet.TopliuneCategoricalArray}
\title{relabelSet for ToplineCategoricalArray}
\usage{
\method{relabelSet}{TopliuneCategoricalArray}(obj, labs)
}
\description{
relabelSet for ToplineCategoricalArray
}
