% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{removeInserts}
\alias{removeInserts}
\title{Remove Inserts}
\usage{
removeInserts(var, theme)
}
\arguments{
\item{var}{A summarized crunch variable}

\item{theme}{A theme object from \link{themeNew}}
}
\description{
Manually drop out Subtotal and Headers from
inserts
}
