% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbook-additions.R
\name{resultsObject}
\alias{resultsObject}
\title{Generic Results Object}
\usage{
resultsObject(x, top = NULL, weighted, body_values, body_labels)
}
\arguments{
\item{x}{A dataset variable}

\item{top}{The top of the results object. NULL by default}

\item{weighted}{Logical. Are these data weighted?}

\item{body_values}{The values to present}

\item{body_labels}{The labels to present}
}
\description{
As \link[crunch]{tabBook} does not provide us with a way to  create summaries
for some variable types we are forced to create an object that bypasses
the reformatVar function. Our goal is to use as much of the
code infrastructure for theming purposes as possible while
allowing the creation of new topline summary objects
}
