% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{roundPropCategorical}
\alias{roundPropCategorical}
\title{Rounds a proportions}
\usage{
roundPropCategorical(data, digits = 0)
}
\arguments{
\item{data}{A numeric vector or CrunchCube object}

\item{digits}{The number of trailing digits}
}
\description{
Round a vector by number of digits. If the rounding
result is greater than 100, rescale to 100
}
