% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{row_data}
\alias{row_data}
\title{Create row data}
\usage{
row_data(data, row, is_crosstabs_array, is_toplines_array, is_base)
}
\arguments{
\item{data}{An object containing data. Either a crosstab array or topline array.}

\item{row}{An integer identifying the row number}

\item{is_crosstabs_array}{Logical, is this a crosstab array?}

\item{is_toplines_array}{Logical, is this a toplines array?}

\item{is_base}{Logical, is this a row of bases?}
}
\description{
Adjust table data to match the desired output
depending on the type of data being presented
}
