% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookLatex.R
\name{scolumnAlign}
\alias{scolumnAlign}
\title{scolumn_fix}
\usage{
scolumnAlign(k, alignment)
}
\arguments{
\item{k}{A data.frame to be printed using \link[kableExtra]{kable}}

\item{alignment}{A string vector of alignments}
}
\description{
Given a data.frame and an alignment vector
create dynamic S-Columns based on character
width
}
