% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catArrayTexAdjustments.R
\name{setCatArrayColNames}
\alias{setCatArrayColNames}
\title{setter for categoricalArray column names}
\usage{
setCatArrayColNames(result, category_aliases)
}
\arguments{
\item{result}{A crosstab result sub-object that is typically passed to \code{\link{writeLatex}}}

\item{category_aliases}{Alternative presentations of the category wording as a character vector.
Must be the same length as the number of categories.}
}
\description{
setter for categoricalArray column names
}
