% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{setResultObj}
\alias{setResultObj}
\title{Generic setter for result object}
\usage{
setResultObj(results, alias, obj)
}
\arguments{
\item{results}{A list of results}

\item{alias}{A character identifying the alias}

\item{obj}{A result object}
}
\description{
Set a result object based on alias
}
