% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabs-sort.R
\name{sortResults}
\alias{sortResults}
\title{Sort Results}
\usage{
sortResults(var, ...)
}
\arguments{
\item{var}{An internal \link{crosstabs} results object.}

\item{...}{Further arguments passed from \link{sortAlias} to type
based \link{sortResults} functions}
}
\description{
An internal function used to apply custom sorting individual questions
in a LateX topline report.
}
