% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabs-sort.R
\name{sortResults_inner}
\alias{sortResults_inner}
\title{sortResults Generic}
\usage{
sortResults_inner(r, descending, alpha, fixed, pin_to_top, pin_to_bottom)
}
\arguments{
\item{r}{The results of a specific banner or Results}
}
\description{
The difference between a topline and a crosstabe from the perspective of tabbook
is that a topline only has a crosstab called "Results" contained within it.
Otherwise, the crosstab object has an embedded summary object per each banner defined
}
