% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{surveyDuration}
\alias{surveyDuration}
\title{Survey Duration}
\usage{
surveyDuration(ds, starttime = NULL, endtime = NULL, outliers = 2.5)
}
\arguments{
\item{ds}{A crunch dataset}

\item{starttime}{A crunch variable containing the starttime}

\item{endtime}{A crunch variable containing the endtime}

\item{outliers}{At what factor of overall mean should outliers be dropped? Defaults to 2.5.}
}
\description{
Calculate survey duration
}
