% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{tab_frame_generate}
\alias{tab_frame_generate}
\title{' tab_frame_generate}
\usage{
tab_frame_generate(default_weight = NULL, vars)
}
\arguments{
\item{default_weight}{A string identifying the default weight}

\item{vars}{A character vector of aliases}
}
\description{
Given a default_weight and a vector of vars, return a tab_frame
that can be used to organize the result of `tabBooks`
}
\details{
Mainly for testing.
}
