% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{tableHeader}
\alias{tableHeader}
\alias{tableHeader.default}
\alias{tableHeader.CrossTabVar}
\alias{tableHeader.ToplineVar}
\alias{tableHeader.ToplineCategoricalArray}
\title{Generate a tableHeader}
\usage{
tableHeader(var, theme)

\method{tableHeader}{default}(var, theme)

\method{tableHeader}{CrossTabVar}(var, theme)

\method{tableHeader}{ToplineVar}(var, theme)

\method{tableHeader}{ToplineCategoricalArray}(var, theme)
}
\arguments{
\item{var}{An object of class ToplineCategoricalArray}

\item{theme}{A theme object from \link{themeNew}}
}
\description{
A passthrough function that creates a table header appropriate to the class
of the data object being passed. Expected classes are:
\itemize{
\item CrossTabVar
\item ToplineVar
\item ToplineCategoricalArray
}

Title indicates whether the title should be displayed, or not (as in the
case of multiple banners displayed underneath each other, the title only
appears on the top one).
Assumes that `\\banner[a-z]{}` macros are defined in the preamble
}
\details{
Importantly, this also controls the relative widths of the columns.
}
