% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{texEscape}
\alias{texEscape}
\title{texEscape}
\usage{
texEscape(string)
}
\arguments{
\item{string}{A string}
}
\description{
A utility function for handling a number of unique
situations that can crop up in tex. (special characters,
British currency, character new line as tex newline)
}
