% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_default}
\alias{theme_default}
\title{Generate default Theme for writeExcel}
\usage{
theme_default(font = getOption("font", default = "Calibri"),
  font_size = getOption("font_size", default = 12),
  font_color = getOption("font_color", default = "black"),
  border_color = getOption("border_color", default = "black"),
  halign = getOption("halign", default = "center"),
  valign = getOption("valign", default = "center"))
}
\arguments{
\item{font}{font font}

\item{font_size}{font_size font_size}

\item{font_color}{font_color font_color}

\item{border_color}{border_color border_color}
}
\description{
\code{theme_default} is the default theme. Users can change base options
}
