% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_latex_default}
\alias{theme_latex_default}
\title{Generate default Theme for writeLatex}
\usage{
theme_latex_default(font = getOption("font", default = "helvet"),
  font_size = getOption("font_size", default = 12))
}
\arguments{
\item{font}{font font}

\item{font_size}{font_size font_size}
}
\description{
\code{theme_latex_default} is the default theme. Users can change base options
}
