% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{themeDefaultExcel}
\alias{themeDefaultExcel}
\title{Generate default Theme for writeExcel}
\usage{
themeDefaultExcel(font = getOption("font", default = "Calibri"),
  font_size = getOption("font_size", default = 12),
  font_color = getOption("font_color", default = "black"),
  valign = getOption("valign", default = "center"))
}
\description{
\code{themeDefaultExcel} is the default theme. Users can change base options
}
