% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{themeDefaultLatex}
\alias{themeDefaultLatex}
\title{Generate default Theme for writeLatex}
\usage{
themeDefaultLatex(font = getOption("font", default = "helvet"),
  font_size = getOption("font_size", default = 12))
}
\arguments{
\item{font}{font font}

\item{font_size}{font_size font_size}
}
\description{
\code{themeDefaultLatex} is the default theme. Users can change base options
}
