% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{themeNew}
\alias{themeNew}
\title{Generate Theme for `writeExcel` or `writeLatex`}
\usage{
themeNew(..., default_theme = themeDefaultExcel())
}
\arguments{
\item{...}{Named arguments that specify the various output types. See the Theme Arguments section below for a complete listing.}

\item{default_theme}{a theme object to use for defaults if you're modifying an already existing themes}
}
\description{
`themeNew` produces themes for `writeExcel` or `writeLatex`.
}
\section{}{
NA
}

