% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplineBase.R
\name{toplineBase}
\alias{toplineBase}
\title{Basic information about a Crunch variable}
\usage{
toplineBase(x)
}
\arguments{
\item{x}{A Crunch variable.}
}
\value{
An object containing basic information about the variable.
}
\description{
\code{toplineBase} returns basic information about a Crunch variable, e.g.
alias, name, type, description.
}

