% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-table.R
\name{toplineTableDef}
\alias{toplineTableDef}
\title{Combine topline data}
\usage{
toplineTableDef(var, tab_definition, header_row, theme)
}
\arguments{
\item{var}{TODO: Identify where this comes from?}

\item{tab_definition}{Defines a longtable or tabular}

\item{header_row}{A different header row depending on if its longtable or tabular}

\item{theme}{A theme object from \link{themeNew}}
}
\description{
Combines topline data to create the Tex
}
