% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackingReports.R
\name{trackingDeclareAvailability}
\alias{trackingDeclareAvailability}
\title{Specify question availability in a tracking report}
\usage{
trackingDeclareAvailability(rebuilt_results, results_available, var, labels)
}
\arguments{
\item{rebuilt_results}{A list of result objects from crunch}

\item{results_available}{A vector identifying in which list elements}

\item{var}{The name of the alias that we are declaring its availability}

\item{labels}{The wave labels}
}
\description{
For each alias, we set an attribute that identifies it's availability
across all the datasets: "general", and "single"
\itemize{
\item "general" means it is available in only some datasets
\item "single" means it is available in exactly one dataset
Because we use subsetting at the list level, "general" and "single"
would follow a typical path that labeling was adjusted appropriately
for presentation in the resulting pdf "single" should act as a simple
passthrough where no additional formatting or manipulation takes place
on the result.
}
}
