% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{usepackage}
\alias{usepackage}
\title{usepackage tex}
\usage{
usepackage(package, ...)
}
\arguments{
\item{package}{A string naming the ctan package to include}

\item{...}{Further arguments passed to the ctan package}
}
\description{
Creates usepackage tex that accepts pacakge
argumentsd
}
