% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{valiases_tabbook_extract}
\alias{valiases_tabbook_extract}
\title{valiases_tabbook_extract}
\usage{
valiases_tabbook_extract(
  is_crosstabs_array,
  crunch_cube,
  cube_variable,
  question_name
)
}
\arguments{
\item{is_crosstabs_array}{A logical identifying if the variable is an array}

\item{crunch_cube}{A sub-cube of a `crunch::tabBook`}

\item{cube_variable}{A sub-cube of a `crunch::tabBook`}
}
\description{
Mainly for testing.
Extracts valiases
}
