% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{validLatexFont}
\alias{validLatexFont}
\title{Validate fonts}
\usage{
validLatexFont(theme_font)
}
\arguments{
\item{theme_font}{A string naming the font specified in the crunchtab theme. Can be one of:
\itemize{
\item bookman
\item charter
\item courier
\item fourier
\item helvet
\item lmodern
\item lmr
\item palatino
\item tgadventor
\item tgbonum
\item tgcursor
\item tgheros
\item tgpagella
\item tgschola
\item tgtermes
\item times
\item utopia
}}
}
\description{
Validate the font passed to a theme object.
}
