% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_api_fixture}
\alias{with_api_fixture}
\title{crunch test wrapper}
\usage{
with_api_fixture(fixture_path, expr)
}
\arguments{
\item{fixture_path}{A full path to fixtures}

\item{expr}{An expression to be run within the api fixture}
}
\description{
Use this to wrap tests that require access to
the crunch api
}
