% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{writeCodebook}
\alias{writeCodebook}
\title{Generate LaTeX CodeBooks}
\usage{
writeCodebook(
  ds,
  theme = themeDefaultLatex(),
  filename = getName(data_summary),
  title = getName(data_summary),
  subtitle = NULL,
  table_of_contents = FALSE,
  sample_desc = NULL,
  field_period = NULL,
  moe = NULL,
  append_text = NULL,
  proportions = FALSE,
  pdf = FALSE,
  open = FALSE,
  logging = FALSE
)
}
\arguments{
\item{ds}{An object of class \code{CrunchDataset}.}

\item{theme}{A theme object (default: `themeDefaultLatex`).}

\item{filename}{character. The name of the output file (without extension).}

\item{title}{An optional title. Defaults to the data summary title.}

\item{subtitle}{An optional character subtitle. Defaults to an empty string.}

\item{table_of_contents}{logical. Should a list of tables be included at the start
of the report Defaults to \code{FALSE}.}

\item{sample_desc}{A character string describing the sample.}

\item{field_period}{A character string describing the field period.}

\item{moe}{An optional numeric margin of error.}

\item{append_text}{An optional character string that, if supplied, will be appended after
the final table. Useful for adding in disclosure information. Defaults to an empty string.}

\item{proportions}{logical. If \code{TRUE} the output report shows proportions,
if \code{FALSE} the report shows counts.}

\item{pdf}{logical. Compile LaTeX using pdflatex? Implemented only on MacOS/Linux.}

\item{open}{logical. If PDF document was produced, open it with
the default application? Only implemented for MacOS.}

\item{logging}{add log messages}
}
\description{
\code{writeCodebook} produces publication-quality LaTeX reports
}
