% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCodeBookLatex.R
\name{writeCodeBookLatex}
\alias{writeCodeBookLatex}
\title{Create a codebook}
\usage{
writeCodeBookLatex(
  ds,
  url = NULL,
  rmd = TRUE,
  pdf = TRUE,
  title = NULL,
  subtitle = NULL,
  table_of_contents = FALSE,
  sample_desc = NULL,
  field_period = NULL,
  preamble = NULL,
  ...
)
}
\arguments{
\item{ds}{A crunch dataset}

\item{url}{A crunch dataset url}

\item{rmd}{Should we create an interim Rmd file? Defaults to TRUE}

\item{pdf}{Should we write directly to pdf? Defaults to TRUE
title = getName(data_summary),}

\item{title}{An optional title. Defaults to the data summary title.}

\item{subtitle}{An optional character subtitle. Defaults to an empty string.}

\item{table_of_contents}{logical. Should a list of tables be included at the start
of the report Defaults to \code{FALSE}.}

\item{sample_desc}{A character string describing the sample.}

\item{field_period}{A character string describing the field period.}

\item{preamble}{A latex string, usually a methodological statement.}

\item{...}{Additional arguments. Unused.}
}
\description{
Create a codebook
}
