% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrong_class_error}
\alias{wrong_class_error}
\title{Assert class}
\usage{
wrong_class_error(value, expected_class, name, null = FALSE)
}
\arguments{
\item{value}{The object}

\item{expected_class}{A character vector of potential classes}

\item{name}{The name of the object}

\item{null}{Logical, identifying if the object can be null}
}
\description{
Function assertion that verifies the
class of an object
}
