% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recontactQuestion.R
\name{as.ToplineCategoricalArray}
\alias{as.ToplineCategoricalArray}
\title{Combine two questions as a categorical array}
\usage{
as.ToplineCategoricalArray(q1, q2, labels = c("Pre", "Post"), weights)
}
\arguments{
\item{q1}{The results object for the first question}

\item{q2}{The results onject second question}

\item{labels}{Two character strings used to describe the pre and post waves}

\item{weights}{A single alias, list, or NULL}
}
\description{
Here we manipulate the tabBook results so that they match the layout
of a categoricalArray, which has the benefit of already having
distinct code to write it to latex.
}
