% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asToplineCategoricalArray.R
\name{catArrayToCategoricals}
\alias{catArrayToCategoricals}
\title{Decompose a categorical array}
\usage{
catArrayToCategoricals(questions, question_alias, labels)
}
\description{
Given two or more waves of a categorical array question, convert them into
categoricals for presentation in a tracking report.
}
