% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebookItem}
\alias{codebookItem}
\alias{codebookItem.default}
\alias{codebookItem.CategoricalVariable}
\alias{codebookItem.CategoricalArrayVariable}
\alias{codebookItem.MultipleResponseVariable}
\alias{codebookItem.NumericVariable}
\alias{codebookItem.TextVariable}
\alias{codebookItem.DatetimeVariable}
\title{Generate a codebook item}
\usage{
codebookItem(x, ...)

\method{codebookItem}{default}(x)

\method{codebookItem}{CategoricalVariable}(x)

\method{codebookItem}{CategoricalArrayVariable}(x)

\method{codebookItem}{MultipleResponseVariable}(x)

\method{codebookItem}{NumericVariable}(x)

\method{codebookItem}{TextVariable}(x)

\method{codebookItem}{DatetimeVariable}(x)
}
\arguments{
\item{x}{An object of one of the types listed}

\item{...}{Additional arguents passed to codebookItem methods}
}
\description{
A passthrough function that creates a table header appropriate to the class
of the data object being passed. Expected classes are:
}
\details{
\itemize{
\item CategoricalVariable
\item CategoricalArrayVariable
\item MultipleResponseVariable
\item NumericVariable
\item TextVariable
\item DatetimeVariable
}

Importantly, this also controls the relative widths of the columns.
}
\section{Methods (by class)}{
\itemize{
\item \code{CategoricalVariable}: Prepares a codebookItem for a CategoricalVariable

\item \code{CategoricalArrayVariable}: Prepares a codebookitem for a CategoricalArrayVariable

\item \code{MultipleResponseVariable}: Prepares a codebookitem for a MultipleResponseVariable

\item \code{NumericVariable}: Prepares a codebookitem for a NumericVariable

\item \code{TextVariable}: Prepares a codebookitem for a TextVariable

\item \code{DatetimeVariable}: Prepares a codebookitem for a DatetimeVariable
}}

