% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookItemBody.R
\name{codeBookItemBody}
\alias{codeBookItemBody}
\alias{codeBookItemBody.default}
\alias{codeBookItemBody.CategoricalVariable}
\alias{codeBookItemBody.CategoricalArrayVariable}
\alias{codeBookItemBody.MultipleResponseVariable}
\alias{codeBookItemBody.DatetimeVariable}
\alias{codeBookItemBody.NumericVariable}
\alias{codeBookItemBody.TextVariable}
\title{codeBookItemBody}
\usage{
codeBookItemBody(x, meta = NULL, ...)

\method{codeBookItemBody}{default}(x, meta, ...)

\method{codeBookItemBody}{CategoricalVariable}(x, meta = NULL, ...)

\method{codeBookItemBody}{CategoricalArrayVariable}(x, meta = NULL, ...)

\method{codeBookItemBody}{MultipleResponseVariable}(x, meta = NULL, ...)

\method{codeBookItemBody}{DatetimeVariable}(x, meta = NULL, ...)

\method{codeBookItemBody}{NumericVariable}(x, meta = NULL, ...)

\method{codeBookItemBody}{TextVariable}(x, meta = NULL, ...)
}
\arguments{
\item{x}{A crunch dataset object}

\item{meta}{A data.frame containing meta data. Must be provided for non-crunch datasets.}

\item{...}{Further arguments, not used.}
}
\description{
Create codebook item body.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default codeBookItemBody.

\item \code{CategoricalVariable}: Creates item body for CategoricalVariable

\item \code{CategoricalArrayVariable}: Creates item body for CategoricalArrayVariable

\item \code{MultipleResponseVariable}: Creates item body for MultipleResponseVariable

\item \code{DatetimeVariable}: Creates item body for DatetimeVariable

\item \code{NumericVariable}: Creates item body for NumericVariable

\item \code{TextVariable}: Creates item body for TextVariable
}}

