% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeBookSummary.R
\name{codeBookSummary}
\alias{codeBookSummary}
\alias{codeBookSummary.default}
\alias{codeBookSummary.factor}
\alias{codeBookSummary.CategoricalVariable}
\alias{codeBookSummary.MultipleResponseVariable}
\alias{codeBookSummary.CategoricalArrayVariable}
\alias{codeBookSummary.NumericVariable}
\alias{codeBookSummary.TextVariable}
\alias{codeBookSummary.DatetimeVariable}
\title{Summarize crunch variables for codeBooks}
\usage{
codeBookSummary(x, meta, ...)

\method{codeBookSummary}{default}(x, meta, ...)

\method{codeBookSummary}{factor}(x, meta, ...)

\method{codeBookSummary}{CategoricalVariable}(x, multiple = FALSE, meta = NULL, ...)

\method{codeBookSummary}{MultipleResponseVariable}(x, meta = NULL, ...)

\method{codeBookSummary}{CategoricalArrayVariable}(x, meta = NULL, ...)

\method{codeBookSummary}{NumericVariable}(x, meta = NULL, ...)

\method{codeBookSummary}{TextVariable}(x, meta = NULL, ...)

\method{codeBookSummary}{DatetimeVariable}(x, meta = NULL, ...)
}
\arguments{
\item{x}{A single variable from a crunch dataset}

\item{meta}{A list containning question meta data}

\item{...}{Additional arguments, unused.}

\item{multiple}{Is this a MultipleResponse or CategoricalArray variable?}
}
\description{
This group of functions creates a summarized data.frame that can be cnoverted
into either a kable or manually glued into a latex.
}
\section{Methods (by class)}{
\itemize{
\item \code{codeBookSummary(default)}: The default, throws out anything that does not
match expected crunch variable classes

\item \code{codeBookSummary(factor)}: Prepares a codeBookSummary data.frame for a factor

\item \code{codeBookSummary(CategoricalVariable)}: Prepares a codeBookSummary data.frame for a
CategoricalVariable

\item \code{codeBookSummary(MultipleResponseVariable)}: Prepares a codeBookSummary data.frame for a
MultipleResponseVariable

\item \code{codeBookSummary(CategoricalArrayVariable)}: Prepares a codeBookSummary data.frame for a
CategoricalArrayVariable

\item \code{codeBookSummary(NumericVariable)}: Prepares a codeBookSummary data.frame for a NumericVariable

\item \code{codeBookSummary(TextVariable)}: Prepares a codeBookSummary data.frame for a TextVariable

\item \code{codeBookSummary(DatetimeVariable)}: Prepares a codeBookSummary data.frame for a DatetimeVaraible

}}
