% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabs.R
\name{crosstabs}
\alias{crosstabs}
\title{Toplines and Crosstabs Summaries}
\usage{
crosstabs(dataset, vars = names(dataset), weight = NULL, banner = NULL,
  codebook = FALSE, title = name(dataset), date = Sys.Date(),
  metadata = NULL)
}
\arguments{
\item{dataset}{A Crunch dataset.}

\item{vars}{An optional vector of aliases of the non-hidden variables that shoulds be used.
Defaults to all non-hidden variables.}

\item{weight}{The alias of a numeric variable that should be used for data weighting.
Defaults to NULL - data is not weighted.}

\item{banner}{An optional object of class \code{Banner} that should be used to generate
a Crosstabs summary. Defaults to NULL - a Toplines summary is produced and returned.}

\item{codebook}{If \code{TRUE}, codebook data summaries are prepared. Defaults to \code{FALSE}.}

\item{title}{An optional title. Defaults to the name of the dataset.}

\item{date}{An optional date. Defaults to the current date.}

\item{metadata}{An optional list with additional metadata that should be added to the summary.}
}
\value{
A Toplines (when no banner is provided) or Crosstabs (when a banner is provided)
summary of the input dataset.
}
\description{
\code{crosstabs} produces Toplines (one-way frequency tables) or Crosstabs (cross tabulations)
summaries of a Cruch dataset.
}
\examples{
\dontrun{
toplines_summary <- crosstabs(crunch_dataset, weight = 'weight')
crosstabs_summary <- crosstabs(crunch_dataset, vars = c('alias1', 'alias2'),
                               weight = 'weight', banner = banner_object)
}
}
