% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchtabs-package.R, R/crunchtabs.R
\docType{package}
\name{crunchtabs-package}
\alias{crunchtabs}
\alias{crunchtabs-package}
\title{crunchtabs: Custom Report Generation for Crunch Datasets}
\description{
In order to generate custom survey reports, this package provides functions for computing 'toplines' (one-way frequency summaries), 'banners' (cross-tabulations) and codebooks of datasets in the Crunch (\url{https://crunch.io/}) web service. Reports can be written in 'PDF' format using 'LaTeX' or in Microsoft Excel '.xlsx' files.

crunchtabs provides tools for generating reports (Toplines and Banners)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Crunch-io/crunchtabs}
  \item Report bugs at \url{https://github.com/Crunch-io/crunchtabs/issues}
}

}
\author{
\strong{Maintainer}: Brandon Bertelsen \email{brandon.bertelsen@yougov.com}

Authors:
\itemize{
  \item Persephone Tsebelis
  \item Kamil Sedrowicz
  \item Neal Richardson \email{neal@crunch.io}
}

}
\keyword{internal}
