% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R, R/codebookLatex.R
\name{kable_styling_defaults}
\alias{kable_styling_defaults}
\title{Defaults for kableExtra}
\usage{
kable_styling_defaults(x, full_width = FALSE, ...)

kable_styling_defaults(x, full_width = FALSE, ...)
}
\arguments{
\item{x}{A kable object}

\item{full_width}{Defaults to TRUE.}

\item{...}{Additional arguments passed to \link[kableExtra]{kable_styling}}
}
\description{
Default styling for kable extra

Default styling for kable extra
}
