% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{reformatLatexResults}
\alias{reformatLatexResults}
\title{reformatCodebookResults <- function(x, digits = 0, reformat = FALSE, round_percentages = FALSE, details = TRUE) {
    UseMethod("reformatCodebookResults", x)
}}
\usage{
reformatLatexResults(data_summary, proportions, theme)
}
\description{
#' @export
reformatCodebookResults.default <- function(x, digits = 0, reformat = FALSE,
    round_percentages = FALSE, details = TRUE) {
    stop("`reformatCodebookResults` generic function doesn't support objects of type:",
        collapse_items(class(x)))
}
}
\details{
#' @export
reformatCodebookResults.ToplineCategoricalGeneral <- function(x, digits = 0, reformat = FALSE,
    round_percentages = FALSE, details = TRUE) {
    x <- setResults(x, reformatResultsGen(x, proportions = FALSE, digits = digits, reformat = reformat,
        round_percentages = round_percentages, details = details),
        proportions = FALSE, details = details)

x <- setResults(x, reformatResultsGen(x, proportions = TRUE, digits = digits, reformat = reformat,
        round_percentages = round_percentages, details = details),
        proportions = TRUE, details = details)
    x
}

#' @export
reformatCodebookResults.ToplineNumeric <- function(x, digits = 0, reformat = FALSE,
    round_percentages = FALSE, details = TRUE) {
    x <- setResults(x, reformatResultsGen(x, digits = digits, details = details),
        details = details)
    x
}

#' @export
reformatCodebookResults.ToplineBase <- function(x, digits = 0, reformat = FALSE,
    round_percentages = FALSE, details = TRUE) {
    x
}
}
