% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabelSet.ToplineVar}
\alias{relabelSet.ToplineVar}
\title{relabel ToplineVar}
\usage{
\method{relabelSet}{ToplineVar}(obj, labs)
}
\arguments{
\item{obj}{A results object for a specific alias}

\item{labs}{A list of labels, typically passed to \link{\code{relabel}}}
}
\description{
Relabel a ToplineVar result object.
}
