% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatResults.R
\name{roundPropCategorical}
\alias{roundPropCategorical}
\title{reformatResults <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    UseMethod("reformatResults", x)
}}
\usage{
roundPropCategorical(data, digits = 0)
}
\description{
#' @export
reformatResults.default <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    stop("`reformatResults` generic function doesn't support objects of type: ",
        collapse_items(class(x)))
}
}
\details{
#' @export
reformatResults.ToplineBase <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}

#' @export
reformatResults.CrossTabVar <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}

#' @export
reformatResults.ToplineNumeric <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}

#' @export
reformatResults.CrossTabBannerVar <- function(x, proportions = TRUE, theme, reformat = TRUE, details = FALSE) {
    reformatResultsGen(x, proportions = proportions, theme = theme,
        reformat = reformat, details = details)
}

#' @importFrom methods is
reformatResultsGen <- function(x, proportions, theme, reformat, details) {
    data <- as.matrix(getResults(x, proportions = proportions, details = details))
    data[is.nan(data)] <- 0
    if (theme$digits > -1 && reformat) {
        if (!proportions || is(x, "ToplineMultipleResponse") || !theme$latex_round_percentages) {
            data[] <- round(data * if (proportions) 100 else 1, theme$digits)
        } else if (is(x, "ToplineCategoricalArray")) {
            data <- t(apply(t(data), 1, roundPropCategorical, theme$digits))
        } else {
            data[] <- roundPropCategorical(data, theme$digits)
        }
    }
    if (proportions && reformat) {
        data[] <- paste0(data, "%")
    }
    data
}
}
