% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabBooks.R
\name{tabBooks}
\alias{tabBooks}
\title{Create a tabBook}
\usage{
tabBooks(dataset, vars, banner, weight = NULL, topline = FALSE)
}
\arguments{
\item{dataset}{A crunch dataset name}

\item{vars}{A character vector of var names that exist within the crunch dataset}

\item{banner}{A banner object from \link{banner}}

\item{weight}{A weighting variable passed to \link{crunch::tabBook}}

\item{topline}{Logical identifying if this is a topline only}
}
\description{
Prepares the actual data for tabulation
}
