% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline.R
\name{topline}
\alias{topline}
\title{Crunch variable summary for toplines reports}
\usage{
topline(var, dataset, weight)
}
\arguments{
\item{var}{A Cruch variable.}

\item{dataset}{A Crunch dataset.}

\item{weight}{A (numeric)Cruch variable that should be used for data weighting.
Defaults to NULL - data is not weighted.}
}
\value{
A toplines summary of a Crunch variable.
}
\description{
\code{topline} returns a toplines summary of a Crunch variable.
}
\examples{
\dontrun{
topline(variable, dataset, weight)
}
}
