% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLatexToplines.R
\name{toplineFooterDef}
\alias{toplineFooterDef}
\title{toplineFooter <- function(x) {
    UseMethod("toplineFooter", x)
}}
\usage{
toplineFooterDef(var_summary)
}
\description{
#' @export
toplineFooter.default <- function(x) {
    toplineFooterDef(is_grid = FALSE)
}
}
\details{
#' @export
toplineFooter.ToplineCategoricalArray <- function(var_summary) {
    toplineFooterDef(is_grid = TRUE)
    
}
}
