% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackingReports.R
\name{tracking_report}
\alias{tracking_report}
\title{Create a multi-dataset tracking report}
\usage{
tracking_report(dataset_list, vars, wave_labels = NULL, weight = NULL)
}
\arguments{
\item{dataset_list}{A list of two or more crunch datasets. Datasets should be 
provided in time order. From oldest to youngest. (i.e, wave 1, wave 2, 
..., wave n)}

\item{vars}{A character vector of question aliases to be included in the report}

\item{wave_labels}{The labels for each wave. Should be of a length that
matches the number of datasets.}

\item{weight}{NULL to accept each dataset's current weight or a single alias
that is available in all datasets as a string.}
}
\description{
This function provides a method for creating a multi-dataset tracking report
}
