% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{writeCodebook}
\alias{writeCodebook}
\title{Generate LaTeX CodeBooks}
\usage{
writeCodebook(...)
}
\arguments{
\item{...}{Additional arguments passed to writeLatx}

\item{ds}{A crunch dataset}
}
\description{
\code{writeCodebook} produces publication-quality LaTeX reports
}
