% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{writeCodeBook}
\alias{writeCodeBook}
\title{Create a codebook}
\usage{
writeCodeBook(ds, url = NULL, rmd = TRUE, pdf = TRUE, ...)
}
\arguments{
\item{ds}{A crunch dataset}

\item{url}{A crunch dataset url}

\item{rmd}{Should we create an interim Rmd file? Defaults to TRUE}

\item{pdf}{Should we write directly to pdf? Defaults to TRUE}

\item{...}{Additional arguments. Unused.}

\item{output}{Defaults to being named after the dataset}
}
\description{
Create a codebook
}
