% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCodeBookLatex.R
\name{writeCodeBookLatex}
\alias{writeCodeBookLatex}
\title{Create a codebook}
\usage{
writeCodeBookLatex(
  ds,
  url = NULL,
  pdf = TRUE,
  title = NULL,
  subtitle = NULL,
  table_of_contents = FALSE,
  sample_desc = NULL,
  field_period = NULL,
  preamble = NULL,
  suppress_zero_counts = FALSE,
  appendix = TRUE,
  logo = NULL,
  position = NULL,
  path = NULL,
  filename = NULL,
  logging = FALSE,
  ...
)
}
\arguments{
\item{ds}{A crunch dataset}

\item{url}{A crunch dataset url}

\item{pdf}{Should we write directly to pdf? Defaults to TRUE.
title = getName(data_summary),}

\item{title}{An optional title. Defaults to the data summary title.}

\item{subtitle}{An optional character subtitle. Defaults to an empty string.}

\item{table_of_contents}{logical. Should a list of tables be included at the start
of the report Defaults to \code{FALSE}.}

\item{sample_desc}{A character string describing the sample.}

\item{field_period}{A character string describing the field period.}

\item{preamble}{A latex string, usually a methodological statement.
LaTeX should be escaped.}

\item{suppress_zero_counts}{Should zero count categories be supressed? Defaults to FALSE.}

\item{appendix}{Should categorical questions with greater than 20 categories be put in an
apppendix? Defaults to TRUE.}

\item{logo}{Default to NULL. A character string one of: yougov or ygblue. Includes the logo
automatically. Also accepts a path to a logo file.}

\item{position}{Defaults to NULL. Identifies the position of the table on the page. Accepts
"c", "l", or "r". Default position is left aligned tables.}

\item{path}{The path to place .tex and .pdf files.}

\item{filename}{A string. The desired basename of the resulting file with no extension
(i.e, "mycodebook")}

\item{logging}{Leave logs in the working directory, defaults to FALSE}

\item{...}{Additional arguments passed to \link[kableExtra]{kable_styling} Unused.}
}
\description{
Create a codebook
}
