% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeExcel.R
\name{writeExcel}
\alias{writeExcel}
\title{Generate Excel Reports: Toplines and Banners}
\usage{
writeExcel(data_summary, filename = NULL, title = getName(data_summary),
  subtitle = NULL, proportions = FALSE, digits = 0,
  table_of_contents = FALSE, logo = NULL, weighted_n = NULL,
  unweighted_n = list(name = "Unweighted N", position = "bottom"),
  show_totals = TRUE, report_desc = NULL, font = "Calibri",
  font_size = 12, one_per_sheet = TRUE, append_text = "",
  banner_vars_split = NULL, row_label_width = 30, reduce_format = FALSE,
  min_base_size = NULL, min_base_label = NULL, show_description = FALSE,
  show_grid_lines = FALSE, open = FALSE, return_data = TRUE,
  logging = FALSE, labels_wrap = list(name = TRUE, description = TRUE,
  row_labels = TRUE, banner_labels = TRUE, column_categories = TRUE),
  first_active_col = 2)
}
\arguments{
\item{data_summary}{An object of class \code{Toplines} or \code{Crosstabs}.}

\item{filename}{The name of the output file (without an extension).}

\item{title}{An optional title. Defaults to the title provided in the summary.}

\item{subtitle}{An optional subtitle. Defaults to an empty string.}

\item{proportions}{logical. If \code{TRUE} the output report shows proportions,
if \code{FALSE} the report shows counts.}

\item{digits}{integer. Number of decimal digits to use for rounding.}

\item{table_of_contents}{logical. Should a list of tables be included at the start
of the report. Defaults to \code{FALSE}.}

\item{logo}{A list of parameters describing the logo:
\itemize{
 \item file - the path to the logo file. Valid file types are: jpeg, png, bmp.
 \item startRow - row coordinate of upper left corner of the image.
 \item startCol - column coordinate of upper left corner of the image.
 \item width - width of figure.
 \item height - height of figure.
 \item units - units of width and height. Can be "in", "cm" or "px".
 \item dpi - image resolution used for conversion between units.
}
Defaults to \code{NULL} - no logo is used.}

\item{weighted_n}{A list of parameters describing the row containing the weighted bases:
\itemize{
 \item name  - row label.
 \item position - row position. Valid values are: "top", "bottom", "both".
}
Defaults to \code{NULL} - the row containing the weighted bases is not printed.}

\item{unweighted_n}{A list of parameters describing the row containing the weighted bases:
\itemize{
 \item name  - row label.
 \item position - row position. Valid values are: "top", "bottom", "both".
}
Defaults to \code{list(name = "Unweighted N", position = "bottom")}.}

\item{show_totals}{logical. If \code{TRUE} a 'Totals' row with column sums is displayed.
Defaults to \code{TRUE}.}

\item{report_desc}{An optional named list of report descriptions.}

\item{font}{A text specifying the font to use. Defaults to 'Calibri'.}

\item{font_size}{Numeric font size points. Defaults to 12 pt font.}

\item{one_per_sheet}{logical. Should every variable be written on a separate sheet?
Defaults to \code{TRUE}.}

\item{append_text}{A text that, if supplied, will be appended after
the final table. Useful for adding in disclosure information.}

\item{banner_vars_split}{the method of splitting banner variables.
Valid values are: NULL, "empty_col".
Defaults to \code{NULL} - no split.}

\item{row_label_width}{width of the first column. Defaults to 30pt.}

\item{reduce_format}{logical. Should the format be reduced?
Defaults to \code{FALSE}.}

\item{min_base_size}{integer. Minimum number of responses for a cross tabulated
category to be displayed in details in a banner report.}

\item{min_base_label}{character. If a number of responses for a
cross tabulated category is less than min_base_size then this text is
used to mask the results. Defaults to \code{NULL} - the values are greyed out.}

\item{show_description}{logical. If \code{TRUE} variables' descriptions are displayed.
Defaults to \code{FALSE}.}

\item{show_grid_lines}{logical. Should the grid lines be shown?}

\item{open}{logical. Should the report, if successfully generated, be opened with
the default application? Defaults to \code{FALSE}.}

\item{return_data}{logical. If \code{TRUE}, a processed data that was used to produce
the report is returned.}

\item{logging}{logical. Should basic information related to the function execution
be printed? Defaults to \code{FALSE}.}

\item{labels_wrap}{list. Specifies which labels should be wrapped:
\itemize{
 \item name  - variable's name.
 \item description - variable's description.
 \item row_labels - row labels.
 \item banner_labels - banner labels
 \item column_categories - banner categories.
}
Defaults to \code{labels_wrap = list(name = TRUE, description = TRUE,
row_labels = TRUE, banner_labels = TRUE, column_categories = TRUE)}.}

\item{first_active_col}{number. First active coulumn. Defaults to \code{2}.}
}
\value{
If \code{return_data} is set to \code{TRUE}, a processed data that was used to produce
the report is returned. Otherwise \code{NULL} is returned.
}
\description{
\code{writeExcel} produces publication-quality Excel reports:
Toplines (one-way frequency tables) or Banners (cross tabulations)
}
\examples{
\dontrun{
# toplines report
toplines_summary <- crosstabs(crunch_dataset, weight = 'weight')
writeExcel(toplines_summary, 'filename')
# crosstabs report
crosstabs_summary <- crosstabs(crunch_dataset, banner = banner_object)
writeExcel(crosstabs_summary, 'filename')
}
}
