% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeExcel.R
\name{writeExcel}
\alias{writeExcel}
\title{Generate Excel Reports: Toplines and Banners}
\usage{
writeExcel(
  data_summary,
  filename = getName(data_summary),
  wb = NULL,
  theme = themeDefaultExcel(),
  title = getName(data_summary),
  subtitle = NULL,
  table_of_contents = FALSE,
  n_or_percent = c("percents", "counts"),
  hypothesis_test = FALSE,
  logging = FALSE,
  save_workbook = TRUE
)
}
\arguments{
\item{data_summary}{An object of class \code{Toplines} or \code{Crosstabs}.}

\item{filename}{character. The name of the output file.}

\item{wb}{An openxlsx Workbook object to add to. Useful for custom front pages.}

\item{theme}{An object of class \code{Theme}.}

\item{title}{character. An optional title. Defaults to the title provided in the summary.}

\item{subtitle}{character. An optional subtitle. Defaults to an empty string.}

\item{table_of_contents}{logical. Should a list of tables be included at the start
of the report. Defaults to \code{FALSE}.}

\item{n_or_percent}{character vector. Should the data be returned as counts or percents?
Valid values are 'counts', 'percents' or both.}

\item{hypothesis_test}{logical. Should hypothesis testing be shown in the tabs?
** Not yet implemented. Defaults to \code{FALSE}.}

\item{logging}{logical. Should basic information related to the function execution
be printed? Defaults to \code{FALSE}.}

\item{save_workbook}{logical. Should the Workbook be saved to an excel file or returned for the
user to continue editing? Defaults to \code{TRUE}.}
}
\value{
If \code{save_workbook} is set to \code{FALSE}, the openxlsx Workbook object is returned
allowing the user to continue editing it before saving. Otherwise the data is returned as it was
provided.
}
\description{
\code{writeExcel} produces publication-quality Excel reports:
Toplines (one-way frequency tables) or Banners (cross tabulations)
}
\examples{
\dontrun{
# toplines report
toplines_summary <- crosstabs(crunch_dataset, weight = "weight")
writeExcel(toplines_summary, "filename")
# crosstabs report
crosstabs_summary <- crosstabs(crunch_dataset, banner = banner_object)
writeExcel(crosstabs_summary, "filename")
}
}
