% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLatex.R
\name{writeLatex}
\alias{writeLatex}
\title{Generate LaTeX Reports: Toplines and Banners}
\usage{
writeLatex(x, filename = NULL, proportions = FALSE, digits = 0,
  title = getName(x), subtitle = NULL, date = Sys.Date(), pdf = FALSE,
  path.to.pdflatex = Sys.which("pdflatex"), open = TRUE,
  returndata = TRUE, table_of_contents = FALSE, moe = NULL,
  headtext = "", foottext = "", sample_desc = "", field_period = "",
  font = "helvet", font_size = NULL, margin = list(top = 0.6, bottom =
  0.6, left = 1, right = 1), append_text = "", longtablewrap = FALSE,
  tableonly = FALSE, landscape = FALSE, pagewidth = ifelse(landscape, 9,
  6.5), min_cell_size = NULL, min_cell_label = NULL, show_totals = TRUE,
  weighted_n = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{Toplines} or \code{Crosstabs}.}

\item{filename}{character. Name of the output file (without extension).}

\item{proportions}{logical. \code{TRUE} returns percentages, \code{FALSE} returns counts.}

\item{digits}{integer. Number of decimal digits to use for rounding.
Defaults to 0.}

\item{title}{An optional title. Defaults to the summary title.}

\item{subtitle}{An optional character subtitle. Defaults to an empty string.}

\item{pdf}{logical. Compile LaTeX using pdflatex? Currently implemented only on Mac/Linux.}

\item{path.to.pdflatex}{Character path to pdflatex.}

\item{open}{logical. If PDF document was produced, open it with
the default application? Currently only implemented for Mac OS X.}

\item{returndata}{logical. If \code{TRUE}, a processed data that was used to produce
the report is returned.}

\item{table_of_contents}{logical. Should a list of tables be included at the start
of the file. Defaults to \code{FALSE}.}

\item{moe}{An optional numeric margin of error.}

\item{headtext}{An optional character string indicating what text should be
placed at the top of continuation tables. 'tbc' is a shortcut for 'To be
continued.'}

\item{foottext}{An optional character string indicating what text should be
placed at the bottom of continuation tables. 'tbc' is a shortcut for
'continued from previous page.'}

\item{sample_desc}{A character string describing the sample.}

\item{field_period}{A character string describing the field period.}

\item{font}{A character string specifying the font package to use in LaTeX.}

\item{font_size}{Font size points for toplines and fontsize name for banners.
Defaults to 12 pt font for toplines and "small" for banners}

\item{margin}{An optional argument to pass to the LaTeX package
\code{geometry}. Default is 'top=.6in, bottom=.6in, left=.5in, right=.5in,
includeheadfoot' for banners and 'top=.6in, bottom=.6in, left=1in,
right=1in, includeheadfoot' for toplines.}

\item{append_text}{An optional character string that, if supplied, will be appended after
the final table. Useful for adding in disclosure information. Defaults to an empty string.}

\item{longtablewrap}{logical. Should LaTeX longtables be allowed to wrap to
a new page? If \code{TRUE}, inserts the 'continued on the next page' breaks.
Defaults to \code{FALSE}.}

\item{tableonly}{logical. If \code{TRUE}, function writes out only the LaTeX
content within the \code{table} environments.}

\item{landscape}{logical. If \code{TRUE}, output pages are in landscape
mode. Defaults to \code{TRUE} for banners and \code{FALSE} for toplines.}

\item{min_cell_size}{integer. Minimum number of responses for a cross tabulated
category to be displayed in details in a banner report.}

\item{min_cell_label}{character. If a number of responses for a
cross tabulated category is less than min_cell_size then this string is
used to mask the results.}

\item{show_totals}{logical. If \code{TRUE} a 'Totals' row with column sums is displayed.
Defaults to \code{TRUE}.}

\item{weighted_n}{logical. Should the total number of responses be weighted?
Defaults to \code{FALSE}.}
}
\value{
If \code{returndata} is set to \code{TRUE}, a processed data that was used to produce
the report is returned. Otherwise \code{NULL} is returned.
}
\description{
\code{writeLatex} produces publication-quality LaTeX reports:
toplines (one-way frequency tables) or banners (cross tabulations).
}
\examples{
\dontrun{
crunch_dataset <- loadDataset('dataset_name')
toplines_summary <- crosstabs(crunch_dataset, weight = 'weight')
writeLatex(toplines_summary, 'filename')
}
}
