% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLatex.R
\name{writeLatex}
\alias{writeLatex}
\title{Generate LaTeX Reports: Toplines and Banners}
\usage{
writeLatex(
  data_summary,
  theme = themeDefaultLatex(),
  filename = getName(data_summary),
  title = getName(data_summary),
  subtitle = NULL,
  table_of_contents = FALSE,
  sample_desc = NULL,
  field_period = NULL,
  moe = NULL,
  append_text = NULL,
  proportions = TRUE,
  pdf = FALSE,
  open = FALSE,
  logging = FALSE
)
}
\arguments{
\item{data_summary}{An object of class \code{Toplines} or \code{Crosstabs}.}

\item{theme}{A theme object (default: `themeDefaultLatex`).}

\item{filename}{character. The name of the output file (without extension).}

\item{title}{An optional title. Defaults to the data summary title.}

\item{subtitle}{An optional character subtitle. Defaults to an empty string.}

\item{table_of_contents}{logical. Should a list of tables be included at the start
of the report Defaults to \code{FALSE}.}

\item{sample_desc}{A character string describing the sample.}

\item{field_period}{A character string describing the field period.}

\item{moe}{An optional numeric margin of error.}

\item{append_text}{An optional character string that, if supplied, will be appended after
the final table. Useful for adding in disclosure information. Defaults to an empty string.}

\item{proportions}{logical. If \code{TRUE} the output report shows proportions,
if \code{FALSE} the report shows counts.}

\item{pdf}{logical. Compile LaTeX using pdflatex? Implemented only on MacOS/Linux.}

\item{open}{logical. If PDF document was produced, open it with
the default application? Only implemented for MacOS.}

\item{logging}{Leave logs in the working directory, defaults to FALSE}
}
\value{
If \code{returndata} is set to \code{TRUE}, a processed data that was used to produce
the report is returned. Otherwise \code{NULL} is returned.
}
\description{
\code{writeLatex} produces publication-quality LaTeX reports:
Toplines (one-way frequency tables) or Banners (cross tabulations).
}
\examples{
\dontrun{
# toplines report
toplines_summary <- crosstabs(crunch_dataset, weight = "weight")
writeLatex(toplines_summary, "filename")
# crosstabs report
crosstabs_summary <- crosstabs(crunch_dataset, banner = banner_object)
writeLatex(crosstabs_summary, "filename")
}
}
